refactor convert api and add tests
