+++ pkg/bundle/plain/v0/convert_test.go
@@ -1,45 +1,45 @@
-package bundle_test
+package v0_test
 
 import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	"github.com/operator-framework/rukpak/pkg/bundle"
 	plainv0 "github.com/operator-framework/rukpak/pkg/bundle/plain/v0"
 	registryv1 "github.com/operator-framework/rukpak/pkg/bundle/registry/v1"
 	"github.com/operator-framework/rukpak/test/testutil"
 )
 
-var _ = Describe("Bundle", func() {
-	Describe("converting a registry+v1 bundle", func() {
+var _ = Describe("Converting from other formats", func() {
+	Describe("Converting from registry+v1", func() {
 		var (
-			srcBundle registryv1.Bundle
-			opts      []func(*plainv0.Bundle)
+			in   registryv1.Bundle
+			opts []func(*plainv0.Bundle)
 		)
 
 		BeforeEach(func() {
-			srcBundle = registryv1.New(testutil.NewRegistryV1FS())
+			in = registryv1.New(testutil.NewRegistryV1FS())
 		})
 
-		var bundle *plainv0.Bundle
+		var out *plainv0.Bundle
 
 		JustBeforeEach(func() {
 			var err error
-			bundle, err = plainv0.FromRegistryV1(srcBundle, opts...)
+			out, err = bundle.Convert(in, opts...)
 			Expect(err).ToNot(HaveOccurred())
 		})
 
 		It("should return a plain+v0 bundle", func() {
-			Expect(bundle).ToNot(BeNil())
+			Expect(out).ToNot(BeNil())
 		})
 
 		It("should have the deployment from the CSV", func() {
 			By("finding all the deployments")
-			deps, err := bundle.Objects(func(obj client.Object) bool {
+			deps, err := out.Objects(func(obj client.Object) bool {
 				_, ok := obj.(*appsv1.Deployment)
 				return ok
 			})
@@ -49,7 +49,7 @@ var _ = Describe("Bundle", func() {
 			Expect(deps[0].GetName()).To(Equal("memcached-operator-controller-manager"))
 
 			By("finding the service account for the deployment")
-			svcAccs, err := bundle.Objects(func(obj client.Object) bool {
+			svcAccs, err := out.Objects(func(obj client.Object) bool {
 				svcAcc, ok := obj.(*corev1.ServiceAccount)
 				if !ok {
 					return false
@@ -64,7 +64,7 @@ var _ = Describe("Bundle", func() {
 
 		It("should have all the RBAC from the CSV", func() {
 			By("extracting the roles")
-			roles, err := bundle.Objects(func(obj client.Object) bool {
+			roles, err := out.Objects(func(obj client.Object) bool {
 				_, ok := obj.(*rbacv1.Role)
 				return ok
 			})
@@ -73,7 +73,7 @@ var _ = Describe("Bundle", func() {
 			Expect(roles).To(HaveLen(0))
 
 			By("extracting the role bindings")
-			roleBindings, err := bundle.Objects(func(obj client.Object) bool {
+			roleBindings, err := out.Objects(func(obj client.Object) bool {
 				_, ok := obj.(*rbacv1.RoleBinding)
 				return ok
 			})
@@ -84,7 +84,7 @@ var _ = Describe("Bundle", func() {
 
 		It("should have all the cluster RBAC from the CSV", func() {
 			By("extracting the cluster roles")
-			roles, err := bundle.Objects(func(obj client.Object) bool {
+			roles, err := out.Objects(func(obj client.Object) bool {
 				_, ok := obj.(*rbacv1.ClusterRole)
 				return ok
 			})
@@ -93,7 +93,7 @@ var _ = Describe("Bundle", func() {
 			Expect(roles).To(HaveLen(1))
 
 			By("extracting the cluster role bindings")
-			roleBindings, err := bundle.Objects(func(obj client.Object) bool {
+			roleBindings, err := out.Objects(func(obj client.Object) bool {
 				_, ok := obj.(*rbacv1.ClusterRoleBinding)
 				return ok
 			})
